+++ src/components/Folder/FolderCard/AllFolderCard.js
@@ -1,6 +1,6 @@
-import { useAsync } from "../../hooks/useAsync";
-import { axiosInstance } from "../../utils/axiosInstance";
-import ReadFolderCard from "./ReadFolderCard";
+import { useAsync } from "../../../hooks/useAsync";
+import { axiosInstance } from "../../../utils/axiosInstance";
+import ReadFolderCard from "../ReadFolderCard";
 
 function AllFolderCard() {
   const getAllFolder = () => axiosInstance.get("users/4/links");
