+++ src/components/Folder/SelectedOption/SelectedOption.css
