+++ src/components/common/SearchBar/SearchBar.js
